% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2sbml.R
\name{convert2sbml}
\alias{convert2sbml}
\title{Writes a SBML file.}
\usage{
convert2sbml(data, outfile)
}
\arguments{
\item{data}{A data.frame with the following mandatory colnames: \code{"ID"},\code{"REACTION"},\code{"GPR"},\code{"LOWER.BOUND"},\code{"UPPER.BOUND"} and \code{"OBJECTIVE"}.}

\item{outfile}{A writable path for the output 'SBML' file to be generated.}
}
\value{
A SBML file.
}
\description{
This function converts a data.frame to a valid SBML file. The Systems Biology Markup Language (SBML) is a representation format, based on XML, for communicating and storing computational models of biological processes. More Info: Encyclopedia of Systems Biology Dubitzky, W., Wolkenhauer, O., Yokota, H., Cho, K.-H. (Eds.) SBML, pp2057-2062 Springer 2013.
}
\details{
This function takes a data.frame as input and convert it to a valid sbmlR object, then the object is writed into the SBML output file.
}
\examples{
 
}
\author{
Daniel Camilo Osorio
}
\keyword{Convert}
\keyword{Metabolic}
\keyword{Reconstruction}
\keyword{SBML}

