% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapReactions.R
\name{mapReactions}
\alias{mapReactions}
\title{Return reactions of a reference data by a selected column}
\usage{
mapReactions(reactionList, referenceData, by, inverse = FALSE)
}
\arguments{
\item{reactionList}{}

\item{referenceData}{}

\item{by}{}

\item{inverse}{}
}
\description{

}
\examples{
# Loading a CSV file
glycolysis <- read.csv2(system.file("extdata", "glycolysisKEGG.csv", package = "minval"))

# Data structure
head(glycolysis)

# Mapping reactions
mapReactions(reactionList = isValidSyntax(glycolysis$REACTION), referenceData = glycolysis, by = "bool")

# Mapping inverse
mapReactions(reactionList = isValidSyntax(glycolysis$REACTION), referenceData = glycolysis, by = "bool", inverse = TRUE)
}
\author{
Daniel Camilo Osorio Hurtado <dcosorioh@unal.edu.co>
}

