\name{minval-package}
\alias{minval}
\docType{package}
\title{
MINimal VALidation for Stoichiometric Reactions
}
\description{
For a given set of stoichiometric reactions, this package evaluates the mass balance and extracts all reactants, products, orphan metabolites, metabolite names and compartments. Also are included some options to check the compound names associated to to Chemical Entities of Biological Interest (ChEBI) database <https://www.ebi.ac.uk/chebi>.}
\details{
\tabular{ll}{
Package: \tab minval\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2016-05-18\cr
License: \tab GPL-2\cr
}
}
\author{
Daniel Osorio, Janneth Gonzalez, Andres Pinzon-Velasco \cr
Maintainer: Daniel Osorio <dcosorioh@unal.edu.co>
}
