% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orphanProducts.R
\name{orphanProducts}
\alias{orphanProducts}
\title{Identify the orphan products of a set of stoichometric reactions}
\usage{
orphanProducts(reactionList, byCompartment = FALSE)
}
\arguments{
\item{reactionList}{A set of stoichiometric reaction with the following format: \code{"H2O[c] + Urea-1-carboxylate[c] <=> 2 CO2[c] + 2 NH3[c]"} Where arrows and plus signs are surrounded by a "space character".
It is also expected that stoichiometry coefficients are surrounded by spaces, (nothe the "2" before the CO2[c] or the NH3[c]).
It also expects arrows to be in the form "\code{=>}" or "\code{<=>}". 
Meaning that arrows like "\code{==>}", "\code{<==>}", "\code{-->}" or "\code{->}" will not be parsed and will lead to errors.}

\item{byCompartment}{A boolean value \code{'TRUE'} or \code{'FALSE'} to indicate if orphan products must be reported by compartment}
}
\description{
This function identifies the orphan products (products that are not consumed in any other reaction or just are involved in one reaction) for a set of stoichometric reactions.
}
\author{
Daniel Camilo Osorio <dcosorioh@unal.edu.co>
}

