% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSBML.R
\name{writeSBML}
\alias{writeSBML}
\title{Write a model in SBML format.}
\usage{
writeSBML(modelData, modelID = "model", outputFile, boundary = "b")
}
\description{
This function converts a data.frame or a modelOrg model to a valid SBML file. The Systems Biology Markup Language (SBML) is a representation format, based on XML, for communicating and storing computational models of biological processes.
}
\author{
Daniel Camilo Osorio <dcosorioh@unal.edu.co>
}
