% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTSV.R
\name{writeTSV}
\alias{writeTSV}
\title{Write a model in TSV format for the 'sybil' R package.}
\usage{
writeTSV(modelData, modelID = "model", outputFile, boundary = "b")
}
\description{
This function converts a data.frame or a modelOrg model to TSV format for the 'sybil' R package.
}
\author{
Daniel Camilo Osorio <dcosorioh@unal.edu.co>
}
