% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls2sbml.R
\name{xls2sbml}
\alias{xls2sbml}
\title{Write a model in SBML format from a EXCEL® spreadsheet}
\usage{
xls2sbml(infile, outfile)
}
\author{
Created by: Tomas Radivoyevitch - Modified and Maintained by: Daniel Camilo Osorio <dcosorioh@unal.edu.co>
}

