% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapReactions.R
\name{mapReactions}
\alias{mapReactions}
\title{Return reactions of a reference data by a selected column}
\usage{
mapReactions(reactionList, referenceData, by, inverse = FALSE)
}
\author{
Daniel Camilo Osorio Hurtado <dcosorioh@unal.edu.co>
}

