% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolites.R
\name{metabolites}
\alias{metabolites}
\title{Identify the list of metabolites for a set of stoichiometric reactions}
\usage{
metabolites(reactionList, woCompartment = FALSE, uniques = TRUE)
}
\arguments{
\item{reactionList}{}

\item{woCompartment}{}

\item{uniques}{}
}
\value{
A list of metabolites for a set of stoichiometric reactions
}
\description{
This function identifies the list of metabolites for a set of stoichiometric reactions. If \code{'woCompartment'} is \code{'TRUE'} compartment label is removed. If \code{'uniques'} is \code{'TRUE'}, list of uniques is returned.
}
\examples{
# Loading data
glycolysis <- read.csv2(system.file("extdata", "glycolysisKEGG.csv", package = "minval"))
glycolysis <- mapReactions(
                           reactionList = isValidSyntax(glycolysis$REACTION),
                           referenceData = glycolysis,
                           by = "bool"
                           )

# Extracting metabolites
metabolites(reactionList = glycolysis$REACTION)

# Extracting metabolites without compartments
metabolites(reactionList = glycolysis$REACTION, 
            woCompartment = TRUE
            )

# Extracting redundant list of metabolites
metabolites(reactionList = glycolysis$REACTION, 
            woCompartment = FALSE,
            uniques = FALSE
            )
}
\author{
Daniel Camilo Osorio <dcosorioh@unal.edu.co>
}

