% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orphanMetabolites.R
\name{orphanMetabolites}
\alias{orphanMetabolites}
\title{Identify the orphan metabolites of a set of stoichiometric reactions}
\usage{
orphanMetabolites(reactionList, actingAs = NULL, byCompartment = FALSE)
}
\arguments{
\item{reactionList}{A set of stoichiometric reaction with the following format:

\code{"H2O[c] + Urea-1-carboxylate[c] <=> 2 CO2[c] + 2 NH3[c]"}

Where arrows and plus signs are surrounded by a "space character".
It is also expected that stoichiometry coefficients are surrounded by spaces, (nothe the "2" before the CO2[c] or the NH3[c]).
It also expects arrows to be in the form "\code{=>}" or "\code{<=>}".
Meaning that arrows like "\code{==>}", "\code{<==>}", "\code{-->}" or "\code{->}" will not be parsed and will lead to errors.}

\item{actingAs}{A text string that specifies the type of metabolite to be returned; only \code{'reactant'} and \code{'product'} are supported.}

\item{byCompartment}{A boolean value \code{'TRUE'} or \code{'FALSE'} to indicate if orphan reactants should be reported by compartment}
}
\value{
If \code{byCompartment == FALSE}, a vector with orphan reactants is returned, in opposite case a list is returned.

If \code{actingAs == 'reactant'}, metabolites not produced in any other reaction or just are involved in one reaction are returned.

If \code{actingAs == 'products'}, metabolites not consumed in any other reaction or just are involved in one reaction are returned.
}
\description{
This function identifies the orphan metabolites (metabolites not produced or not consumed in any other reaction or just involved in one reaction) for a set of stoichometric reactions.
}
\examples{
# Loading data
glycolysisKEGG <- read.csv2(system.file("extdata", "glycolysisKEGG.csv", package = "minval"), sep = "\\t")

# Extracting orphan reactants
orphanMetabolites(reactionList = glycolysis$REACTION, actingAs = "reactant")

# Extracting orphan products by compartment
orphanMetabolites(reactionList = glycolysis$REACTION, actingAs = "product", byCompartment = TRUE)
}
\author{
Daniel Camilo Osorio <dcosorioh@unal.edu.co>
}
